v2: rewrite EventChan to prevent inotify instance leak
EventChan is meant to monitor the memory.events file of a v2 cgroup and feed
events into a channel. It does this by setting up an inotify monitor for the
events file, and blocking on read of the inotify instance. There is no way to
interrupt/abort this goroutine. Contrary to common belief, close(fd) does not
unblock readers currently blocked in ther kernel, this is a userspace thing
that go does in it's i/o code.

What this results in is an inotify instance leak in containerd if a container
exits for any reason, and is restarted. As long as containerd keeps running,
this will leak both an inotify instance and a goroutine blocked on the read.
The way to fix this is to add a pipe and epoll instance, use epoll to monitor
the inotify fd and pipe, and close the write end of the pipe to unblock the
whole thing. That is already implemented in github.com/fsnotify/fsnotify, so
rewrite the inotify handling to use that package.

fsnotify exposes an error and event channel, and we can select on those an
handle events. The select allows us to return an additional channel to the
user, the 'quitCh', that a user can close to signal that we should stop
processing events.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>