v2: monitor cgroup events and quit when cgroup becomes empty
The contract expected from containerd is that when a cgroup is removed, we will
stop monitoring for OOM events. If we don't follow that contract, a goroutine
and associated inotify instance will always be leaked when a container exits,
as long as containerd continues running.

Try to detect cgroup deletion by monitoring cgroup.events as well. When the
cgroup gets unpopulated, it means the last process exited and we stop
monitoring for OOM.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>